fix(at): clear bonds on AT+PAIR=reset, remove PURGE, fix %llu logs
- AT+PAIR=reset: add ble_clear_bonds() + settings_save() calls
  before storage_format_card() to persist the bond deletion
  through the reboot. Includes <zephyr/settings/settings.h>.

- Remove AT+PURGE and AT+PURGEABLE commands (unused; PURGEABLE
  used %llu which newlib-nano printf doesn't support).

- Fix 3 remaining %llu/%lld log sites with %u + unsigned casts
  (newlib-nano prints nothing for long-long formats).

+++ applications/clip/src/audio.c
@@ -656,7 +656,7 @@ void audio_recording_thread(void *p1, void *p2, void *p3)
                 mic_power_off();
 
                 is_paused = true;
-                LOG_INF("Recording paused at %llu sec", stats.recording_time_ms / 1000);
+                LOG_INF("Recording paused at %u sec", (unsigned int)(stats.recording_time_ms / 1000));
 
                 /* Wait for resume signal */
                 while (is_paused && !stop_requested) {
